+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Collocated the time-display and time-sorting settings for files to reduce confusion if users _display_ the modification time but sort using the creation time, or vice versa.
 - Improved the layouting of the display settings tab.
 - The context menu is now a custom one, making the experience more seamless.
+- If you change the display settings for the editor, the editor will now also remove rendered elements that you do not wish to be rendered anymore.
 
 ## Under the Hood